Moved WeatherForecast to shared service
