+++ BlazorClientTest/Shared/Interfaces/IWeatherForecastService.cs
@@ -1,0 +1,10 @@
+﻿using BlazorClientTest.Shared.Models;
+using System.Collections.Generic;
+
+namespace BlazorClientTest.Shared.Interfaces
+{
+    public interface IWeatherForecastService
+    {
+        IEnumerable<WeatherForecast> Get();
+    }
+}
