+++ BlazorClientTest/Shared/Models/WeatherForecast.cs
@@ -1,17 +1,12 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace BlazorClientTest.Shared
+namespace BlazorClientTest.Shared.Models
 {
     public class WeatherForecast
     {
         public DateTime Date { get; set; }
-
         public int TemperatureC { get; set; }
-
         public string Summary { get; set; }
-
         public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
     }
 }
