+++ BlazorClientTest/Shared/Services/WeatherForecastService.cs
@@ -1,0 +1,28 @@
+﻿using BlazorClientTest.Shared.Interfaces;
+using BlazorClientTest.Shared.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace BlazorClientTest.Shared.Services
+{
+    public class WeatherForecastService : IWeatherForecastService
+    {
+        private static readonly string[] Summaries = new[]
+        {
+            "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
+        };
+
+        public IEnumerable<WeatherForecast> Get()
+        {
+            var rng = new Random();
+            return Enumerable.Range(1, 5).Select(index => new WeatherForecast
+            {
+                Date = DateTime.Now.AddDays(index),
+                TemperatureC = rng.Next(-20, 55),
+                Summary = Summaries[rng.Next(Summaries.Length)]
+            })
+            .ToArray();
+        }
+    }
+}
